Dispacher: Parsing: Elevator door opening
Implements filtering of repeated positional event in order to parse
probable instances of the elevator doors opening.

This does introduce side effects in the elevator function as it
previously was relyant on a 'DoorState' structure which has been
altered to match the other possible events.

This probably should have been done in a separate commit but I'm
tired and wrote this comment instead.

Big Note: This only works when opening the doors while the elveator
is level as it's based on the presumtion that the positional events
will have the same position.

Signed-off-by: Rasmus Linusson <rasmus@linusson.org>