Dispatcher: logic: Raniking system
Update to the elevator ranking system.

For each elevator the stop queue list is iterated until a suitable
position is presented. Each stop is weighten more than the total
travel distance as the elevator has to stop to allow passengers
to enter and exit the cabin.

The elevator with the lowest scrore is thereby most suited to
serivce the floor. This however clompletly disregards any
power consumtion factor in scheduling the elvators.